v0.2.6

+++ src/main/java/stu/sll/main/common/uiEvent/product/ProductDelete.java
@@ -1,17 +1,17 @@
 package stu.sll.main.common.uiEvent.product;
 
-import stu.sll.main.common.dialog.DialogErr;
-import stu.sll.main.common.dialog.DialogMsg;
+import stu.sll.main.common.util.DialogErr;
+import stu.sll.main.common.util.DialogMsg;
 import stu.sll.main.database.operations.product.DeleteProduct;
 
 import javax.swing.*;
 import java.sql.SQLException;
 
 /**
- * @BelongsProject: Market
- * @BelongsPackage: stu.sll.main.common.uiEvent.product
- * @Author: ZhouChenyi
- * @Description: Product Deelete
+ * @Title: ProductDelete
+ * @Author: Chenlvin
+ * @Package: stu.sll.main.common.uiEvent.product
+ * @Description: Delete product
  */
 
 public class ProductDelete {
@@ -19,13 +19,18 @@ public class ProductDelete {
         if(row == -1) { // 未选择行
             new DialogMsg("提示","请先选择需要操作的对象");
         } else {
-            // 删除操作
-            int selectedPid = (int) ProductTable.getValueAt(row, 0);
-            try {
-                DeleteProduct.deleteProduct(selectedPid);
-                new ProductSearch("名称", "", ProductTable);
-            } catch (SQLException ex) {
-                new DialogErr(ex.getMessage());
+            int flag = JOptionPane.showConfirmDialog(new JFrame(), "确认删除商品 \"" + ProductTable.getValueAt(row, 1) + "\" 吗？",
+                    "提示", JOptionPane.YES_NO_OPTION,
+                    JOptionPane.INFORMATION_MESSAGE);
+            if(JOptionPane.YES_OPTION == flag){
+                // 点击确认，执行删除操作
+                int selectedPid = (int) ProductTable.getValueAt(row, 0);
+                try {
+                    DeleteProduct.deleteProduct(selectedPid);
+                    new ProductSearch("名称", "", ProductTable);
+                } catch (SQLException ex) {
+                    new DialogErr(ex.getMessage());
+                }
             }
         }
     }
